+++ comic_dl/__init__.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 
 # import sites
-# # import manga_eden
 # # import comic_dl
 # # import globalFunctions
 # # from comic_dl import ComicDL