+++ Changelog.md
@@ -134,3 +134,4 @@
 - Fix for #310 [2022.04.20]
 - Added support for Qiremanhua [Fix for #193] [2022.04.23]
 - New host chapmanganato [#331] [2023.01.08]
+- Cleanup mangaEden [#323] [2023.01.08]